pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voice-agent-server"
version = "1.0.0"
description = "Voice AI assistant server with LiveKit Agents"
requires-python = ">=3.9"

dependencies = [
    "livekit-agents[silero,turn-detector,elevenlabs,openai,deepgram]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "playwright>=1.40.0",
    "httpx>=0.25.0",
    "google-auth>=2.3.0",
    "google-api-python-client>=2.0.0",
    "google-auth-oauthlib>=0.4.6",
    "firebase-admin>=6.0.0",
    "pyrebase4>=4.7.0",
    "setuptools>=80.9.0",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-dir]
"" = "."

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
